GWIGD-16 GWIGD-17 GWIGD-18 # resolve #comment Implement a mini map
* Implement a mini map which gives the user an overview about the game world and indicates where he/she is.

* Improve the scrolling

* Add a "blocked" and "water" map array which needed to determine where we can construct buildings and where we can have a burial at see.

* Remove the test map and add the desert map.
I have had to rearranged the map in the editor to fit into the game. The map has now just one background layer, called "terra", layers for blocked tiles and for water and one for the entities. I assume the last is not necessary because we put add the entities in the game to the map and this will be sprites. Anyway, the layer is there and maybe we will need it, otherwise I could delete the layer.
I added a new tileset to indicate water and blocked tiles (It won't be rendered in game). The single tiles have properties (W-tile = water and X-tile = blocked) which I need to fill the water and blocked map.

Its also important to stick on this layer config (terra = Layer0, water = Layer 1, blocked = layer 2, ...) to get the tile id from it.

* Make some refactoring to get rid of the fot/fit variables in every state because the were defined and initialized already in the view class.
Same goes for game and container variables in some states.